refactor folder structure for tests
