+++ test/shared/mocks.ts
@@ -8,7 +8,7 @@ import { contractAdminAddress } from "./constants";
 export async function deployMockAirBroFactory(deployer: Signer): Promise<MockContract> {
   const airBroFactory: MockContract = await waffle.deployMockContract(deployer, AirBroFactory.abi);
 
-  await airBroFactory.mock.admin.returns("0xF4b5bFB92dD4E6D529476bCab28A65bb6B32EFb3");
+  await airBroFactory.mock.admin.returns(contractAdminAddress);
 
   return airBroFactory;
 }
@@ -16,7 +16,7 @@ export async function deployMockAirBroFactory(deployer: Signer): Promise<MockCon
 export async function deployMockAirBroFactory1155Holder(deployer: Signer): Promise<MockContract> {
   const airBroFactory1155Holder: MockContract = await waffle.deployMockContract(deployer, AirBroFactory1155Holder.abi);
 
-  await airBroFactory1155Holder.mock.admin.returns("0xF4b5bFB92dD4E6D529476bCab28A65bb6B32EFb3");
+  await airBroFactory1155Holder.mock.admin.returns(contractAdminAddress);
 
   return airBroFactory1155Holder;
 }
